+++ src/ssh/azext_ssh/_params.py
@@ -27,7 +27,7 @@ def load_arguments(self, _):
                    help=('This is an internal argument. This argument is used by Azure Portal to provide a one click '
                          'SSH login experience in Cloud shell.'),
                    deprecate_info=c.deprecate(hide=True), action='store_true')
-        c.argument('arc_proxy_folder', options_list=['--arc-proxy-folder'],
+        c.argument('ssh_proxy_folder', options_list=['--ssh-proxy-folder'],
                    help=('Path to the folder where the arc proxy should be saved. '
                          'Default to .clientsshproxy folder in user\'s home directory if not provided.'))
         c.positional('ssh_args', nargs='*', help='Additional arguments passed to OpenSSH')
@@ -50,7 +50,7 @@ def load_arguments(self, _):
         c.argument('resource_type', options_list=['--resource-type'],
                    help='Resource type should be either Microsoft.Compute or Microsoft.HybridCompute')
         c.argument('cert_file', options_list=['--certificate-file', '-c'], help='Path to certificate file')
-        c.argument('arc_proxy_folder', options_list=['--arc-proxy-folder'],
+        c.argument('ssh_proxy_folder', options_list=['--ssh-proxy-folder'],
                    help=('Path to the folder where the arc proxy should be saved. '
                          'Default to .clientsshproxy folder in user\'s home directory if not provided.'))
 
@@ -75,7 +75,7 @@ def load_arguments(self, _):
                    help=('This is an internal argument. This argument is used by Azure Portal to provide a one click '
                          'SSH login experience in Cloud shell.'),
                    deprecate_info=c.deprecate(hide=True), action='store_true')
-        c.argument('arc_proxy_folder', options_list=['--arc-proxy-folder'],
+        c.argument('ssh_proxy_folder', options_list=['--ssh-proxy-folder'],
                    help=('Path to the folder where the arc proxy should be saved. '
                          'Default to .clientsshproxy folder in user\'s home directory if not provided.'))
         c.positional('ssh_args', nargs='*', help='Additional arguments passed to OpenSSH')